relocate hoststats impl to types
